Update to fits_editor.py
Now the fits editor not only adds/edits the header keys we discussed today but writes the .FIT file to a .fits file organized as following

<script_directory>/yyyy/mm/dd/yyyymmdd_zzz_xxxx.fits

The code should be sufficiently commented, if you have any questions let me know. I used os.path.join() for creating the paths so I'm expecting it to correctly create paths no matter what system it is used on, I'll try it on a virtual Ubuntu image tonight and a virtual OSX image if I can find the one I used to have.

I guess we might want to separate this into 2 different scripts later, one for editing the original values and one for saving it as a .fits file organized by the directories.