mplot1 and mplot3 updates
Changed mplot1 to now have the optional argument for step size for -f
flag. so -f fb fe fs. mplot3 can how take multiple flags so mplot3.py -x
400 -y 450 will show and save two images named slice_x.fits and
slice_y.fits respectively. The function that slices the cube cslice()
takes a header, data, a flag string, and a flag value and will slice it
and return data and a header so it can be used stand-alone from the rest
of the script.