Progress on Pictures
Pictures update automatically... some of the time. It refreshes when the image it's pulling is open in a tab but for some reason it doesn't update when it isn't open. This may or may not be an issue the server will have,  but I don't think it will because the server will be more directly connected to the image than my laptop.
In other news, the picture is centered properly without having the weird behaviors it was having previously